+++ scripts/ci/pylint_todo.txt
@@ -19,7 +19,7 @@
 ./airflow/contrib/hooks/opsgenie_alert_hook.py
 ./airflow/providers/apache/pinot/hooks/pinot.py
 ./airflow/contrib/hooks/qubole_check_hook.py
-./airflow/contrib/hooks/sagemaker_hook.py
+./airflow/providers/amazon/aws/hooks/sagemaker.py
 ./airflow/contrib/hooks/segment_hook.py
 ./airflow/contrib/hooks/slack_webhook_hook.py
 ./airflow/contrib/hooks/snowflake_hook.py
